+++ lib/SILOptimizer/ARC/ARCMatchingSet.cpp
@@ -11,10 +11,8 @@
 //===----------------------------------------------------------------------===//
 
 #define DEBUG_TYPE "arc-sequence-opts"
-#include "GlobalARCPairingAnalysis.h"
 #include "RefCountState.h"
-#include "GlobalARCSequenceDataflow.h"
-#include "GlobalLoopARCSequenceDataflow.h"
+#include "ARCMatchingSet.h"
 #include "swift/Basic/BlotMapVector.h"
 #include "swift/Basic/Fallthrough.h"
 #include "swift/SIL/SILBuilder.h"
@@ -39,80 +37,11 @@ using namespace swift;
 //                          ARC Matching Set Builder
 //===----------------------------------------------------------------------===//
 
-namespace {
-
-struct MatchingSetFlags {
-  bool KnownSafe;
-  bool Partial;
-};
-static_assert(std::is_pod<MatchingSetFlags>::value,
-              "MatchingSetFlags should be a pod.");
-
-struct ARCMatchingSetBuilder {
-  using TDMapTy = BlotMapVector<SILInstruction *, TopDownRefCountState>;
-  using BUMapTy = BlotMapVector<SILInstruction *, BottomUpRefCountState>;
-
-  TDMapTy &TDMap;
-  BUMapTy &BUMap;
-
-  llvm::SmallVector<SILInstruction *, 8> NewIncrements;
-  llvm::SmallVector<SILInstruction *, 8> NewDecrements;
-  bool MatchedPair;
-  ARCMatchingSet MatchSet;
-  bool PtrIsGuaranteedArg;
-
-  RCIdentityFunctionInfo *RCIA;
-
-public:
-  ARCMatchingSetBuilder(TDMapTy &TDMap, BUMapTy &BUMap,
-                        RCIdentityFunctionInfo *RCIA)
-    : TDMap(TDMap), BUMap(BUMap), MatchedPair(false),
-      PtrIsGuaranteedArg(false), RCIA(RCIA) {}
-
-  void init(SILInstruction *Inst) {
-    clear();
-    MatchSet.Ptr = RCIA->getRCIdentityRoot(Inst->getOperand(0));
-
-    // If we have a function argument that is guaranteed, set the guaranteed
-    // flag so we know that it is always known safe.
-    if (auto *A = dyn_cast<SILArgument>(MatchSet.Ptr)) {
-      if (A->isFunctionArg()) {
-        auto C = A->getParameterInfo().getConvention();
-        PtrIsGuaranteedArg = C == ParameterConvention::Direct_Guaranteed;
-      }
-    }
-    NewIncrements.push_back(Inst);
-  }
-
-  void clear() {
-    MatchSet.clear();
-    MatchedPair = false;
-    NewIncrements.clear();
-    NewDecrements.clear();
-  }
-
-  bool matchUpIncDecSetsForPtr();
-
-  // We only allow for get result when this object is invalidated via a move.
-  ARCMatchingSet &getResult() {
-    return MatchSet;
-  }
-
-  bool matchedPair() const { return MatchedPair; }
-
-private:
-  /// Returns .Some(MatchingSetFlags) on success and .None on failure.
-  Optional<MatchingSetFlags> matchIncrementsToDecrements();
-  Optional<MatchingSetFlags> matchDecrementsToIncrements();
-};
-
-} // end anonymous namespace
-
 /// Match retains to releases and return whether or not all of the releases
 /// were known safe.
 Optional<MatchingSetFlags>
 ARCMatchingSetBuilder::matchIncrementsToDecrements() {
-  MatchingSetFlags Flags = { true, false };
+  MatchingSetFlags Flags = {true, false};
 
   // For each increment in our list of new increments.
   //
@@ -124,7 +53,7 @@ ARCMatchingSetBuilder::matchIncrementsToDecrements() {
     assert(BURefCountState != BUMap.end() && "Catch this on debug builds.");
     if (BURefCountState == BUMap.end()) {
       DEBUG(llvm::dbgs() << "        FAILURE! Could not find state for "
-            "increment!\n");
+                            "increment!\n");
       return None;
     }
 
@@ -134,11 +63,12 @@ ARCMatchingSetBuilder::matchIncrementsToDecrements() {
     // nothing we can pair it with implying we should skip it.
     if (!(*BURefCountState)->second.isTrackingRefCountInst()) {
       DEBUG(llvm::dbgs() << "    SKIPPING INCREMENT! State not tracking any "
-            "instruction.\n");
+                            "instruction.\n");
       continue;
     }
 
-    // We need to be known safe over all increments/decrements we are matching up
+    // We need to be known safe over all increments/decrements we are matching
+    // up
     // to ignore insertion points.
     bool BUIsKnownSafe = (*BURefCountState)->second.isKnownSafe();
     DEBUG(llvm::dbgs() << "        KNOWNSAFE: "
@@ -157,8 +87,10 @@ ARCMatchingSetBuilder::matchIncrementsToDecrements() {
       SILInstruction *Decrement = DecIter;
       DEBUG(llvm::dbgs() << "        Decrement: " << *Decrement);
 
-      // Now grab the increment matched up with the decrement from the bottom up map.
-      // If we can't find it, bail we can't match this increment up with anything.
+      // Now grab the increment matched up with the decrement from the bottom up
+      // map.
+      // If we can't find it, bail we can't match this increment up with
+      // anything.
       auto TDRefCountState = TDMap.find(Decrement);
       if (TDRefCountState == TDMap.end()) {
         DEBUG(llvm::dbgs() << "            FAILURE! Could not find state for "
@@ -168,7 +100,8 @@ ARCMatchingSetBuilder::matchIncrementsToDecrements() {
       DEBUG(llvm::dbgs() << "            SUCCESS! Found state for "
                             "decrement.\n");
 
-      // Make sure the increment we are looking at is also matched to our decrement.
+      // Make sure the increment we are looking at is also matched to our
+      // decrement.
       // Otherwise bail.
       if (!(*TDRefCountState)->second.isTrackingRefCountInst() ||
           !(*TDRefCountState)->second.containsInstruction(Increment)) {
@@ -199,7 +132,7 @@ ARCMatchingSetBuilder::matchIncrementsToDecrements() {
 
 Optional<MatchingSetFlags>
 ARCMatchingSetBuilder::matchDecrementsToIncrements() {
-  MatchingSetFlags Flags = { true, false };
+  MatchingSetFlags Flags = {true, false};
 
   // For each increment in our list of new increments.
   //
@@ -211,7 +144,7 @@ ARCMatchingSetBuilder::matchDecrementsToIncrements() {
     assert(TDRefCountState != TDMap.end() && "Catch this on debug builds.");
     if (TDRefCountState == TDMap.end()) {
       DEBUG(llvm::dbgs() << "        FAILURE! Could not find state for "
-            "increment!\n");
+                            "increment!\n");
       return None;
     }
 
@@ -221,11 +154,12 @@ ARCMatchingSetBuilder::matchDecrementsToIncrements() {
     // nothing we can pair it with implying we should skip it.
     if (!(*TDRefCountState)->second.isTrackingRefCountInst()) {
       DEBUG(llvm::dbgs() << "    SKIPPING DECREMENT! State not tracking any "
-            "instruction.\n");
+                            "instruction.\n");
       continue;
     }
 
-    // We need to be known safe over all increments/decrements we are matching up
+    // We need to be known safe over all increments/decrements we are matching
+    // up
     // to ignore insertion points.
     bool TDIsKnownSafe = (*TDRefCountState)->second.isKnownSafe();
     DEBUG(llvm::dbgs() << "        KNOWNSAFE: "
@@ -244,8 +178,10 @@ ARCMatchingSetBuilder::matchDecrementsToIncrements() {
       SILInstruction *Increment = IncIter;
       DEBUG(llvm::dbgs() << "        Increment: " << *Increment);
 
-      // Now grab the increment matched up with the decrement from the bottom up map.
-      // If we can't find it, bail we can't match this increment up with anything.
+      // Now grab the increment matched up with the decrement from the bottom up
+      // map.
+      // If we can't find it, bail we can't match this increment up with
+      // anything.
       auto BURefCountState = BUMap.find(Increment);
       if (BURefCountState == BUMap.end()) {
         DEBUG(llvm::dbgs() << "            FAILURE! Could not find state for "
@@ -256,7 +192,8 @@ ARCMatchingSetBuilder::matchDecrementsToIncrements() {
       DEBUG(
           llvm::dbgs() << "            SUCCESS! Found state for increment.\n");
 
-      // Make sure the increment we are looking at is also matched to our decrement.
+      // Make sure the increment we are looking at is also matched to our
+      // decrement.
       // Otherwise bail.
       if (!(*BURefCountState)->second.isTrackingRefCountInst() ||
           !(*BURefCountState)->second.containsInstruction(Decrement)) {
@@ -269,7 +206,8 @@ ARCMatchingSetBuilder::matchDecrementsToIncrements() {
       // Add the decrement to the decrement to move set. If we don't insert
       // anything, just continue.
       if (!MatchSet.Increments.insert(Increment)) {
-        DEBUG(llvm::dbgs() << "    SKIPPING! Already processed this increment.\n");
+        DEBUG(llvm::dbgs()
+              << "    SKIPPING! Already processed this increment.\n");
         continue;
       }
 
@@ -295,7 +233,8 @@ bool ARCMatchingSetBuilder::matchUpIncDecSetsForPtr() {
 
   while (true) {
     DEBUG(llvm::dbgs() << "Attempting to match up increments -> decrements:\n");
-    // For each increment in our list of new increments, attempt to match them up
+    // For each increment in our list of new increments, attempt to match them
+    // up
     // with decrements and gather the insertion points of the decrements.
     auto Result = matchIncrementsToDecrements();
     if (!Result) {
@@ -344,7 +283,6 @@ bool ARCMatchingSetBuilder::matchUpIncDecSetsForPtr() {
     MatchSet.DecrementInsertPts.clear();
   } else {
     DEBUG(llvm::dbgs() << "NOT UNCONDITIONALLY SAFE!\n");
-
   }
 
   bool HaveIncInsertPts = !MatchSet.IncrementInsertPts.empty();
@@ -371,96 +309,4 @@ bool ARCMatchingSetBuilder::matchUpIncDecSetsForPtr() {
   // Success!
   DEBUG(llvm::dbgs() << "SUCCESS! We can move, remove things.\n");
   return true;
-}
-
-//===----------------------------------------------------------------------===//
-//                            ARC Pairing Context
-//===----------------------------------------------------------------------===//
-
-bool ARCPairingContext::performMatching(CodeMotionOrDeleteCallback &Callback) {
-  bool MatchedPair = false;
-
-  DEBUG(llvm::dbgs() << "**** Computing ARC Matching Sets for " << F.getName()
-                     << " ****\n");
-
-  /// For each increment that we matched to a decrement, try to match it to a
-  /// decrement -> increment pair.
-  for (auto Pair : IncToDecStateMap) {
-    if (!Pair.hasValue())
-      continue;
-
-    SILInstruction *Increment = Pair->first;
-    if (!Increment)
-      continue; // blotted
-
-    DEBUG(llvm::dbgs() << "Constructing Matching Set For: " << *Increment);
-    ARCMatchingSetBuilder Builder(DecToIncStateMap, IncToDecStateMap, RCIA);
-    Builder.init(Increment);
-    if (Builder.matchUpIncDecSetsForPtr()) {
-      MatchedPair |= Builder.matchedPair();
-      auto &Set = Builder.getResult();
-      for (auto *I : Set.Increments)
-        IncToDecStateMap.blot(I);
-      for (auto *I : Set.Decrements)
-        DecToIncStateMap.blot(I);
-
-      // Add the Set to the callback. *NOTE* No instruction destruction can
-      // happen here since we may remove instructions that are insertion points
-      // for other instructions.
-      Callback.processMatchingSet(Set);
-    }
-  }
-
-  // Then finalize the callback. This is the only place instructions can be
-  // deleted.
-  Callback.finalize();
-  return MatchedPair;
-}
-
-//===----------------------------------------------------------------------===//
-//                                  Loop ARC
-//===----------------------------------------------------------------------===//
-
-void LoopARCPairingContext::runOnLoop(SILLoop *L) {
-  auto *Region = LRFI->getRegion(L);
-  if (processRegion(Region, false, false)) {
-    // We do not recompute for now since we only look at the top function level
-    // for post dominating releases.
-    processRegion(Region, true, false);
-  }
-
-  // Now that we have finished processing the loop, summarize the loop.
-  Evaluator.summarizeLoop(Region);
-}
-
-void LoopARCPairingContext::runOnFunction(SILFunction *F) {
-  if (processRegion(LRFI->getTopLevelRegion(), false, false)) {
-    // We recompute the final post dom release since we may have moved the final
-    // post dominated releases.
-    processRegion(LRFI->getTopLevelRegion(), true, true);
-  }
-}
-
-bool LoopARCPairingContext::processRegion(const LoopRegion *Region,
-                                          bool FreezePostDomReleases,
-                                          bool RecomputePostDomReleases) {
-  bool MadeChange = false;
-  bool NestingDetected = Evaluator.runOnLoop(Region, FreezePostDomReleases,
-                                             RecomputePostDomReleases);
-  bool MatchedPair = Context.performMatching(Callback);
-  MadeChange |= MatchedPair;
-  Evaluator.clearLoopState(Region);
-  Context.DecToIncStateMap.clear();
-  Context.IncToDecStateMap.clear();
-
-  while (NestingDetected && MatchedPair) {
-    NestingDetected = Evaluator.runOnLoop(Region, FreezePostDomReleases, false);
-    MatchedPair = Context.performMatching(Callback);
-    MadeChange |= MatchedPair;
-    Evaluator.clearLoopState(Region);
-    Context.DecToIncStateMap.clear();
-    Context.IncToDecStateMap.clear();
-  }
-
-  return MadeChange;
 }
