+++ lib/SILOptimizer/ARC/ARCSequenceOpts.cpp
@@ -12,7 +12,7 @@
 
 #define DEBUG_TYPE "arc-sequence-opts"
 #include "swift/SILOptimizer/PassManager/Passes.h"
-#include "GlobalARCPairingAnalysis.h"
+#include "ARCSequenceOpts.h"
 #include "swift/Basic/Fallthrough.h"
 #include "swift/SIL/SILBuilder.h"
 #include "swift/SIL/SILVisitor.h"
@@ -40,6 +40,98 @@ STATISTIC(NumRefCountOpsMoved, "Total number of increments moved");
 STATISTIC(NumRefCountOpsRemoved, "Total number of increments removed");
 
 llvm::cl::opt<bool> EnableLoopARC("enable-loop-arc", llvm::cl::init(true));
+
+//===----------------------------------------------------------------------===//
+//                            ARC Pairing Context
+//===----------------------------------------------------------------------===//
+
+bool ARCPairingContext::performMatching(CodeMotionOrDeleteCallback &Callback) {
+  bool MatchedPair = false;
+
+  DEBUG(llvm::dbgs() << "**** Computing ARC Matching Sets for " << F.getName()
+                     << " ****\n");
+
+  /// For each increment that we matched to a decrement, try to match it to a
+  /// decrement -> increment pair.
+  for (auto Pair : IncToDecStateMap) {
+    if (!Pair.hasValue())
+      continue;
+
+    SILInstruction *Increment = Pair->first;
+    if (!Increment)
+      continue; // blotted
+
+    DEBUG(llvm::dbgs() << "Constructing Matching Set For: " << *Increment);
+    ARCMatchingSetBuilder Builder(DecToIncStateMap, IncToDecStateMap, RCIA);
+    Builder.init(Increment);
+    if (Builder.matchUpIncDecSetsForPtr()) {
+      MatchedPair |= Builder.matchedPair();
+      auto &Set = Builder.getResult();
+      for (auto *I : Set.Increments)
+        IncToDecStateMap.blot(I);
+      for (auto *I : Set.Decrements)
+        DecToIncStateMap.blot(I);
+
+      // Add the Set to the callback. *NOTE* No instruction destruction can
+      // happen here since we may remove instructions that are insertion points
+      // for other instructions.
+      Callback.processMatchingSet(Set);
+    }
+  }
+
+  // Then finalize the callback. This is the only place instructions can be
+  // deleted.
+  Callback.finalize();
+  return MatchedPair;
+}
+
+//===----------------------------------------------------------------------===//
+//                                  Loop ARC
+//===----------------------------------------------------------------------===//
+
+void LoopARCPairingContext::runOnLoop(SILLoop *L) {
+  auto *Region = LRFI->getRegion(L);
+  if (processRegion(Region, false, false)) {
+    // We do not recompute for now since we only look at the top function level
+    // for post dominating releases.
+    processRegion(Region, true, false);
+  }
+
+  // Now that we have finished processing the loop, summarize the loop.
+  Evaluator.summarizeLoop(Region);
+}
+
+void LoopARCPairingContext::runOnFunction(SILFunction *F) {
+  if (processRegion(LRFI->getTopLevelRegion(), false, false)) {
+    // We recompute the final post dom release since we may have moved the final
+    // post dominated releases.
+    processRegion(LRFI->getTopLevelRegion(), true, true);
+  }
+}
+
+bool LoopARCPairingContext::processRegion(const LoopRegion *Region,
+                                          bool FreezePostDomReleases,
+                                          bool RecomputePostDomReleases) {
+  bool MadeChange = false;
+  bool NestingDetected = Evaluator.runOnLoop(Region, FreezePostDomReleases,
+                                             RecomputePostDomReleases);
+  bool MatchedPair = Context.performMatching(Callback);
+  MadeChange |= MatchedPair;
+  Evaluator.clearLoopState(Region);
+  Context.DecToIncStateMap.clear();
+  Context.IncToDecStateMap.clear();
+
+  while (NestingDetected && MatchedPair) {
+    NestingDetected = Evaluator.runOnLoop(Region, FreezePostDomReleases, false);
+    MatchedPair = Context.performMatching(Callback);
+    MadeChange |= MatchedPair;
+    Evaluator.clearLoopState(Region);
+    Context.DecToIncStateMap.clear();
+    Context.IncToDecStateMap.clear();
+  }
+
+  return MadeChange;
+}
 
 //===----------------------------------------------------------------------===//
 //                                Code Motion
